Adjusted rehydration and examiner records to function with DSP
